Major Update
Changed the structure of the files. The Greplib file is now a command
line interface that is used to call parts of Greplib from the command
line with arguments. The old contents of the Greplib file have been
moved to Greplib.Process. Timers were optimized, output was reformatted,
lots of unnecessary code was removed.